Initial commit for re-design of Pysos.
Pysos, as it was, was a large hack. Poorly written, no consistency...a disaster.

This is the first commit for the re-design of pysos into a proper python-y utility.
Most modules have been recoded and are now proper modules. There is still work to be done on
the rhevm, storage, lspci modules but that will come soon enough.

Long term goal with this re-design is to also incorporate a webUI for visualization of data
beyond the pretty print outs of pysos currently.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>