Incorrect parsing causes stack traces
In bios.py we were parsing for the 'BIOS' section rather than 'BIOS Information' - which could lead to parsing of the 'BIOS Language' seciton which is obviously not going to contain the right data. Fixed to point to the right section.

In network.py we were not returning the dictionary properly. Corrected.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>