Fixed divide by zero chance and bad total memory rounding
If no swap space is defined, we would previously hit a divide-by-zero and fail. Here we skip swap reports if there is no swap (but report there is no swap).

Also, removed an old test that was causing total RAM to be displayed as +1GB. Removed that and replaced the round() with a math.ceil()

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>