Fixed incorrect parsing of proc/net/dev values
Depending on the OS there may or may not be a space between the device name and the number of received bytes
in proc/net/dev. Previously we did not account for this so if there *was* a space, we would be off by one column
when assigning the values from the file to the dictionary. We now check and account for the space (or lack thereof)

Also made all dictionary values integers to start, rather than converting them later. Adjusted format of netdev output to compensate for this.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>