Fixed incorrect location of lspci file and efficiency improvements
For some reason we were looking in the wrong location for the lspci file, so we'd always report it was missing when it was not. Now pointing at the correct location.

Also we now only get the raw lspci file information once, during intilization instead of randomly throughout the various methods.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>