Fix device reporting header
In displayLspciInfo() we previously were using the devType key as the device type header, which worked for most devices. However, some SCSI device names could break the parsing and give some nasty long string for the devType key. We now use the chkType variable for this instead of relying on the parser for every device under the sun.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>