Correcting inadvertant case changing and reducing string for bond mode
Was unintentially formatting the interface names in self.devList to lowercase - so if the device had a mix of upper and lower case characters, we would fail to collect information on it. Removed this case change.

Additionally, reducing the length of the bonding mode string by removing the name in () that /proc/net/bonding/bondX reports.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>