Virt updates to ps changes, correcting ps command attr assignment, expanding virt module
Updated the virt module to account for ps's new use of modules instead of dictionaries.
Also adding the getHyperInfo() method for virt module to return a hyper object with appropriate hypervisor info based on the hypervisor type.

Further, corrected bug in ps that was truncating the command attr to the first whole string in the command string.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>