Add docker and Atomic Host module
Adding in docker and Atomic Host parsing. Both are obtained by the '--docker' option and if the system is running a kubernetes configuration, that will be reported as well.

Note that this only supported v1beta3 for kubernetes at this point since kube version 0.15 deprecates v1beta1 and v1beta3 is the rc for the v1 api.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>